fix: DFScalar accepts Int without explicit Int64() cast (#205) (#214)
Replace the bare `Variant` alias for `DFScalar` with a thin struct that
adds an `@implicit __init__(value: Int)` overload normalising `Int` to
`Int64` at construction time.  All four existing typed arms keep their
`@implicit` decorators so dict/list literals still coerce without
explicit wrappers.

Callers can now write `{"key": 42}` instead of `{"key": Int64(42)}`.
Update the from_records tests to use plain integer literals.

Closes #205

+++ tests/test_dataframe.mojo
@@ -418,8 +418,8 @@ fn test_itertuples_name() raises:
 
 
 fn test_from_records_basic() raises:
-    var row0: Dict[String, DFScalar] = {"a": Int64(1), "b": Int64(10)}
-    var row1: Dict[String, DFScalar] = {"a": Int64(2), "b": Int64(20)}
+    var row0: Dict[String, DFScalar] = {"a": 1, "b": 10}
+    var row1: Dict[String, DFScalar] = {"a": 2, "b": 20}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
     records.append(row1^)
@@ -438,7 +438,7 @@ fn test_from_records_empty() raises:
 
 
 fn test_from_records_columns_param() raises:
-    var row0: Dict[String, DFScalar] = {"a": Int64(1), "b": Int64(2), "c": Int64(3)}
+    var row0: Dict[String, DFScalar] = {"a": 1, "b": 2, "c": 3}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
     var cols = List[String]()
@@ -451,8 +451,8 @@ fn test_from_records_columns_param() raises:
 
 
 fn test_from_records_mixed_types() raises:
-    var row0: Dict[String, DFScalar] = {"i": Int64(42), "s": "hello"}
-    var row1: Dict[String, DFScalar] = {"i": Int64(7), "s": "world"}
+    var row0: Dict[String, DFScalar] = {"i": 42, "s": "hello"}
+    var row1: Dict[String, DFScalar] = {"i": 7, "s": "world"}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
     records.append(row1^)
@@ -463,7 +463,7 @@ fn test_from_records_mixed_types() raises:
 
 fn test_from_records_int_float_mixed() raises:
     # First row has Int64, second row has Float64 — column should be promoted to float64
-    var row0: Dict[String, DFScalar] = {"x": Int64(1)}
+    var row0: Dict[String, DFScalar] = {"x": 1}
     var row1: Dict[String, DFScalar] = {"x": Float64(2.5)}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
@@ -481,7 +481,7 @@ fn test_from_records_int_float_mixed() raises:
 fn test_from_records_bool_int_mixed() raises:
     # First row has Bool, second row has Int64 — column should be promoted to int64
     var row0: Dict[String, DFScalar] = {"y": True}
-    var row1: Dict[String, DFScalar] = {"y": Int64(42)}
+    var row1: Dict[String, DFScalar] = {"y": 42}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
     records.append(row1^)
@@ -496,8 +496,8 @@ fn test_from_records_bool_int_mixed() raises:
 
 
 fn test_from_records_missing_key() raises:
-    var row0: Dict[String, DFScalar] = {"a": Int64(1), "b": Int64(10)}
-    var row1: Dict[String, DFScalar] = {"a": Int64(2)}
+    var row0: Dict[String, DFScalar] = {"a": 1, "b": 10}
+    var row1: Dict[String, DFScalar] = {"a": 2}
     # "b" is missing in row1
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
@@ -513,7 +513,7 @@ fn test_from_records_missing_key() raises:
 fn test_from_records_column_order_deterministic() raises:
     # Column names should be sorted alphabetically when `columns` is not provided,
     # so the result is deterministic regardless of Dict iteration order.
-    var row0: Dict[String, DFScalar] = {"z": Int64(3), "a": Int64(1), "m": Int64(2)}
+    var row0: Dict[String, DFScalar] = {"z": 3, "a": 1, "m": 2}
     var records = List[Dict[String, DFScalar]]()
     records.append(row0^)
     var df = DataFrame.from_records(records)
